Refactor linkcheck steps into separate files
